Example code for random forest models
Example of the python code used for the random forest models built as part of an ongoing RUSVM Center 4 Research Project assessing predictors of student success.  This example code is attributed to the aspect of the project described in the manuscript "The Use of Random Forest Machine Learning Models for Veterinary Student Recruitment and Retention: Predicting Academic Success and Identification of At Risk- Students".